StimuliNet/flownet2/pyproject.toml
[tool.poetry]
name = "flownet2"
version = "0.1.0"
description = "FlowNet2"
authors = ["Debajyoti Raychaudhuri <draychaudhuri1086@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6.8"
tensorflow-gpu = "2.2.0"
tensorflow-addons = "0.11.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
